feat(infieldprogress): add new component
