+++ components/infieldprogresscircle/README.md
@@ -1,0 +1,7 @@
+# @spectrum-css/infieldprogresscircle
+
+> The Spectrum CSS infield progress circle component
+
+This package is part of the [Spectrum CSS project](https://github.com/adobe/spectrum-css).
+
+See the [Spectrum CSS documentation](https://opensource.adobe.com/spectrum-css/infieldprogresscircle).
