+++ components/infieldprogresscircle/metadata/mods.md
@@ -1,0 +1,3 @@
+| Modifiable custom properties         |
+| ------------------------------------ |
+| `--mod-progress-cirlce-stroke-width` |
