+++ components/progresscircle/index.css
@@ -87,28 +87,6 @@ governing permissions and limitations under the License.
 	}
 }
 
-.spectrum-ProgressCircle--infield {
-	--spectrum-progress-circle-thickness: var(--spectrum-progress-circle-thickness-small);
-	--spectrum-progress-circle-size: var(--spectrum-in-field-progress-circle-size-100);
-	padding-block: var(--spectrum-in-field-progress-circle-edge-to-fill);
-
-	&.spectrum-ProgressCircle--sizeS {
-		--spectrum-progress-circle-size: var(--spectrum-in-field-progress-circle-size-75);
-	}
-
-	&.spectrum-ProgressCircle--sizeL {
-		--spectrum-progress-circle-size: var(--spectrum-in-field-progress-circle-size-200);
-	}
-
-	&.spectrum-ProgressCircle--sizeXL {
-		--spectrum-progress-circle-size: var(--spectrum-in-field-progress-circle-size-300);
-	}
-
-	.spectrum-ProgressCircle-fill {
-		stroke-linecap: square;
-	}
-}
-
 /* windows high contrast mode */
 @media (forced-colors: active) {
 	.spectrum-ProgressCircle {
