+++ components/progresscircle/stories/template.js
@@ -17,7 +17,6 @@ export const Template = ({
 	size = "m",
 	staticColor,
 	isIndeterminate = false,
-	isInField = false,
 	value,
 }) => {
 
@@ -31,7 +30,6 @@ export const Template = ({
 			class=${classMap({
 				[rootClass]: true,
 				[`${rootClass}--indeterminate`]: isIndeterminate,
-				[`${rootClass}--infield`]: isInField,
 				[`${rootClass}--static${capitalize(staticColor)}`]: typeof staticColor !== "undefined",
 				[`${rootClass}--size${size?.toUpperCase()}`]: typeof size !== "undefined" && size !== "m",
 				...customClasses.reduce((a, c) => ({ ...a, [c]: true }), {}),
